update to 19w06a and fabric 0.2.0
