+++ gradle.properties
@@ -1,8 +1,8 @@
 #General
-mc_version=19w05a
-yarn_mappings=19w05a.6
-loader_version=0.3.4.104
-fabric_version=0.1.5.87
+mc_version=19w06a
+yarn_mappings=19w06a.3
+loader_version=0.3.5.106
+fabric_version=0.2.0.92
 
 #Publishing
 mod_version = 0.2.0-SNAPSHOT
